Bug fixes: pawn always could move two. Moves not prevented from check.
* Fixed a bug where any pawn always had the option to move two, even if
it had already moved.
* Fixed a bug where kings were apparently able (and willing, it seems)
to put themselves into check.  I can't say much for how "willing" they
still are, but it should not be registering as a legal move anymore.
* Uncovered castling bug with new unit test.  White king-side castling
is not being listed as an available move when all conditions are met.
Should fix and check the other castling moves as well.